change uint16 to unsigned

+++ Cote/Renderer/API/VertexAttributeLayout.cpp
@@ -2,8 +2,8 @@
 
 using namespace cote::graphic;
 
-void cote::graphic::VertexAttributeLayout::push(VertexAttributeIndex index, uint16_t type, uint16_t count, uint16_t size, bool normalized)
+void cote::graphic::VertexAttributeLayout::push(VertexAttributeIndex index, unsigned type, unsigned count, unsigned size, bool normalized)
 {
-	m_attributes.push_back({ index,type,count,m_stride,size });
+	m_attributes.push_back({ (unsigned)index,type,count,m_stride,size });
 	m_stride += size*count;
 }
